+++ Development/DeviceTestView.swift
@@ -1,5 +1,5 @@
 import SwiftUI
-#if canImport(Device) // since this is needed in XCode but is unavailable in Playgrounds
+#if canImport(Device) // since this is needed in XCode but is unavailable in Playgrounds.
 import Device
 #endif
 
@@ -38,9 +38,15 @@ struct TimeClockView: View {
             Text("Current time: \(time.formatted(date: .long, time: .complete))")
             if let battery = Device.current.battery {
                 Text("Battery Info: \(battery.description)")
-                BatteryView(battery: battery, fontSize: 80)
+                HStack {
+                    BatteryView(battery: battery, fontSize: 80)
+                    Image(systemName: Device.current.symbolName)
+                        .font(.system(size: 80))
+                }
             } else {
                 Text("No Battery")
+                Image(systemName: Device.current.symbolName)
+                    .font(.system(size: 80))
             }
         }
         .onReceive(timer, perform: { _ in
@@ -50,10 +56,20 @@ struct TimeClockView: View {
     }
 }
 
+struct StackedLabelStyle: LabelStyle {
+    func makeBody(configuration: Configuration) -> some View {
+        VStack {
+            configuration.icon.font(.title2)
+            configuration.title.font(.caption2)
+        }
+    }
+}
+
 struct Placard: View {
     @State var color = Color.gray
     var body: some View {
-        if #available(macCatalyst 17.0, iOS 17.0, *) {
+        // if platform missing, this causes crash.  Also, this needs the stroke(_:lineWidth:antialiased version which is only iOS 17+ to work in previews.
+        if #available(iOS 17.0, macOS 14.0, macCatalyst 17.0, tvOS 17.0, watchOS 10.0, visionOS 1.0, *) {
             return RoundedRectangle(cornerRadius: 10)
                 .stroke(.primary, lineWidth: 3)
                 .fill(color)
@@ -67,64 +83,100 @@ struct Placard: View {
 
 struct TestCard: View {
     @State var label = "Unknown"
-    @State var visible = true
+    @State var highlighted = true
     @State var color = Color.gray
+    @State var symbolName = String.symbolUnknownEnvironment
     var body: some View {
-        Placard(color: visible ? color : .clear)
+        Placard(color: highlighted ? color : .clear)
             .overlay {
-                Text(label)
+                Label(label, systemImage: symbolName)
                     .font(.caption)
+                    .symbolRenderingMode(highlighted ? .hierarchical : .monochrome)
             }
     }
 }
 
 public struct DeviceTestView: View {
-    public var body: some View {
-        VStack {
-            Group { // so not more than 7 items
-                Text("Kudit/Device v\(Bundle.main.infoDictionary?["CFBundleShortVersionString"] ?? "Unknown")")
-                TimeClockView()
-                Text("Current device: \(Device.current.description)")
-                Text("Identifier: \(Device.current.identifier)")
-                Text("Device Name: \(Device.current.name ?? "nil")")
-                Text("System Name: \(Device.current.systemName ?? "nil")")
-            }
-            Group {
-                HStack {
-                    TestCard(
-                        label: "Preview",
-                        visible: Device.current.isPreview,
-                        color: .orange)
-                    TestCard(
-                        label: "Playground",
-                        visible: Device.current.isPlayground,
-                        color: .pink)
-                    TestCard(
-                        label: "Simulator",
-                        visible: Device.current.isSimulator,
-                        color: .blue)
-                    TestCard(
-                        label: "Real Device",
-                        visible: Device.current.isRealDevice,
-                        color: .green)
-                }
-                .frame(height: 44)
-                HStack {
-                    VStack {
-                        BatteryTestView(useSystemColors: true, fontSize: 40)
-                    }
-                    VStack {
-                        ForEach(Device.Idiom.allCases) { idiom in
-                            TestCard(label: idiom.description, visible: Device.current.idiom == idiom, color: Device.current.idiom.color)
-                        }
-                    }
-                    VStack {
-                        BatteryTestView(includePercent: false, fontSize: 40)
-                    }
-                }
-            }
-            .padding()
-            Spacer()
+    public var version: String {
+        if let version = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String {
+            return version
+        }
+        return "Unknown"        
+    }
+    public var idiomList: some View {
+        ForEach(Device.Idiom.allCases) { idiom in
+            TestCard(
+                label: idiom.description,
+                highlighted: Device.current.idiom == idiom,
+                color: Device.current.idiom.color,
+                symbolName: idiom.symbolName)
         }
     }
+    
+    public var body: some View {
+        ScrollView {
+            VStack {
+                Group { // so not more than 7 items
+                    Text("Kudit/Device v\(version)")
+                    TimeClockView()
+                    Text("Current device: \(Device.current.description)")
+                    Text("Identifier: \(Device.current.identifier)")
+                    Text("Device Name: \(Device.current.name ?? "nil")")
+                    Text("System Name: \(Device.current.systemName ?? "nil")")
+                }
+                Group {
+                    HStack {
+                        //                    TestCard(label: "TEST", highlighted: true, color: .yellow, symbolName: "star.fill")
+                        TestCard(
+                            label: "Preview",
+                            highlighted: Device.current.isPreview,
+                            color: .orange,
+                            symbolName: .symbolPreview
+                        )
+                        TestCard(
+                            label: "Playground",
+                            highlighted: Device.current.isPlayground,
+                            color: .pink,
+                            symbolName: .symbolPlayground)
+                        TestCard(
+                            label: "Simulator",
+                            highlighted: Device.current.isSimulator,
+                            color: .blue,
+                            symbolName: .symbolSimulator)
+                        TestCard(
+                            label: "Real Device",
+                            highlighted: Device.current.isRealDevice,
+                            color: .green,
+                            symbolName: .symbolRealDevice)
+                        if [.mac, .vision].contains(Device.current.idiom) {
+                            TestCard(
+                                label: "Designed for iPad",
+                                highlighted: Device.current.isDesignedForiPad,
+                                color: .purple,
+                                symbolName: .symbolDesignedForiPad)
+                        }
+                    }
+                    .labelStyle(StackedLabelStyle())
+                    .frame(height: 60)
+                    HStack {
+                        VStack {
+                            BatteryTestView(useSystemColors: true, fontSize: 40)
+                        }
+                        VStack {
+                            idiomList
+                        }
+                        VStack {
+                            BatteryTestView(includePercent: false, fontSize: 40)
+                        }
+                    }
+                }
+                .padding()
+                Spacer()
+            }
+        }
+    }
+}
+
+#Preview {
+    DeviceTestView()
 }