+++ Tests/DeviceTests.swift
@@ -5,8 +5,9 @@ import XCTest
 final class DeviceTests: XCTestCase {
     func testDeviceOutput() async throws {
         // TODO: Clearly this needs work.  If someone wants to replace this with good tests, it would be very much appreciated!
-        print(Device.current.description)
-        XCTAssert(Device.current.description == Device(identifier: "iPad16,2").description, "Device test")
+        print("Current description: \(Device.current.name ?? "Unknown")")
+        print("Created description: \(Device(identifier: "iPhone16,1").name)")
+        XCTAssert(Device.current.name == Device(identifier: "iPhone16,1").name, "Device test")
         // This is an example of a functional test case.
         // Use XCTAssert and related functions to verify your tests produce the correct
         // results.