+++ Development/Resources/CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+v1.1.0 3/20/2024 Added symbols for device idiom icons and various build modes.  Fixed TestView Any to String conversion issue.  Added dynamicIsland configuration parameter.  Added in missing iPhone 14 Pro Max and updated several screen sizes and some wrong support IDs.  Added isMini flag for HomePods.  Added Mac forms. Created test for (Designed for iPad) mode vs native mode.  Added XCode test project so that we can build on visionOS and tvOS and watchOS and various mac targets.  visionOS (Designed for iPad) properly reports the new isDesignedForiPad flag when using a compatibility app.  Added scroll view so visible on watchOS or in iPhone landscape.  Added tvOS, visionOS, macOS, and watchOS icons.
 v1.0.11 3/11/2024 Fixed BatteryView to make sure it updates when DeviceBattery updates.  Fixed so macOS reports charging instead of full when charging.
 v1.0.10 3/11/2024 Updated documentation.  Changed phone test color to red since gray is not very obvious.  Re-worked BatteryView so that it can be scaled.
 v1.0.9 3/11/2024 Fixed so that BatteryView actually uses the initialization parameters (formerly, would not because they were marked @State).
@@ -16,23 +17,29 @@ v1.0.0 2/16/2024 Initial Project based off of DeviceKit but designed to be more 
 ## Bugs to fix:
 Known issues that need to be addressed.
 
-[ ] Fix so that previews consistently work in Swift Playgrounds.
+[ ] Fix so that previews consistently work in Xcode & Swift Playgrounds.
     - Does not show #Preview("Battery") in Playgrounds on macOS or on iPad Swift Playgrounds.  App Preview is shown.
     - All Previews work in Xcode.
-[ ] Running on macOS (Designed for iPad) reports OS as iPadOS and returns an iPad Pro identifier instead of a Mac identifier and battery information is incorrect.  Need to test in visionOS (Designed for iPad) to see if there are similar issues in that configuration.
+[ ] Designed for iPad running on macOS has all appearance of being an actual iPad and battery level does not work.  Need help on this edge case (or use macCatalyst or macOS development).  Running on macOS (Designed for iPad) reports OS as iPadOS and returns an iPad Pro identifier instead of a Mac identifier and battery information is incorrect.
+[ ] Need help getting identifier when buildling for macOS (not catalyst)
+[ ] Building from Playground (not using Xcode project), Designed for iPad doesn't report properly but identifier is correct (systemName reports iPadOS) - same when buildling for Mac Catalyst.  Buildling from the Xcode project Designed for iPad does propertly report isDesignedForiPad but the battery indicator and device is wrong.  Buildling for Mac Catalyst does propertly report device and battery.
 
 ## Roadmap:
 Planned features and anticipated API changes.  If you want to contribute, this is a great place to start.
 
-[ ] Add weight option for primary text (title originally was bold and now that we're doing font size, add weight parameter)
+[ ] Add weight option for battery text/symbol (title originally was bold and now that we're doing font size, add weight parameter)
 [ ] Add macOS devices (ongoing).
 [ ] Add new devices (ongoing).
 [ ] Improve test app user interface/layout.
 [ ] Add tests to complete code coverage.
 [ ] Create tests in such a way that they can be run in Previews and in the example App (may require a project dependency).
+[ ] Support landscape on iPhone for test view and use size classes to re-layout.
+[ ] For devices that support multiple orientations, automatically apply current orientation to the symbol name for current device.
 
 ## Proposals:
 This is where proposals can be discussed for potential movement to the roadmap.
 
 [ ] Change Battery from Protocol to the DeviceBattery code and have MockBattery be a subclass that overrides functions?
 [ ] Create better way of scaling battery so that it can scale with dynamic type?
+[ ] Add a way to check that privacy checks have been added when using APIs that need privacy permissions (have a configuration flag that needs to be set to ensure that privacy flags have been set).
+[ ] Convert SF Symbol names to an enum to allow version checking?